Removed the obsolete code/files
